Fixed big message crash when calling Update (fixes #54)

+++ LemonUI/Scaleform/BigMessage.cs
@@ -301,7 +301,7 @@ namespace LemonUI.Scaleform
                     CallFunction(function, Title, Message, Rank);
                     break;
                 case MessageType.Weapon:
-                    CallFunction(function, Title, Message, weaponHash);
+                    CallFunction(function, Title, Message, (int)weaponHash);
                     break;
                 default:
                     CallFunction(function, Title, Message);
